Add files via upload
This is the Week-2 Assignment which covers the Pandas basic Concepts.
As a beginner, I took help from Googling and StackOverflow to complete Assignments.
I was so excited while doing my first ever Data Science Assignment.

Week-1:
in the first week, I was having only Quiz of 12 MCQs question. in that assignment, I scored 91.66% marks :) 
the assignment was all about introduction to Python for Data Science.